Extract invocation matching into `InvocationShape`
This moves all logic from `MethodCall` that is related to invocation
matching into a dedicated new type, `InvocationShape`. This seems use-
ful mainly because of two reasons:

 1. The invocation matching becomes more easily testable once it is
    better isolated.

 2. Verification methods such as `mock.Verify(expression)` no longer
    need to instantiate a transient `MethodCall` (which at its heart
    really represents a setup) in order to match invocations. They
    will be able to use the much lighter-weight `InvocationShape` in-
    stead.
